# api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "text/csv"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation for Store Manager API",
    "title": "Store Manager API",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/api/status": {
      "get": {
        "description": "Get status of an job",
        "produces": [
          "application/json"
        ],
        "tags": [
          "StoreManager"
        ],
        "operationId": "status",
        "parameters": [
          {
            "description": "Job ID",
            "name": "jobid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success response",
            "schema": {
              "$ref": "#/definitions/StatusResponse"
            }
          },
          "400": {
            "description": "error response",
            "schema": {
              "$ref": "#/definitions/StatusResponse"
            }
          },
          "500": {
            "description": "error response",
            "schema": {
              "$ref": "#/definitions/StatusResponse"
            }
          }
        }
      }
    },
    "/api/submit": {
      "post": {
        "description": "Creates jobs",
        "produces": [
          "application/json"
        ],
        "tags": [
          "StoreManager"
        ],
        "operationId": "submit",
        "parameters": [
          {
            "description": "Job request",
            "name": "requestBody",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SubmitRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "success response",
            "schema": {
              "$ref": "#/definitions/SubmitResponse"
            }
          },
          "400": {
            "description": "error response",
            "schema": {
              "$ref": "#/definitions/SubmitResponse"
            }
          },
          "500": {
            "description": "error response",
            "schema": {
              "$ref": "#/definitions/SubmitResponse"
            }
          }
        }
      }
    },
    "/api/visits": {
      "get": {
        "description": "Creates jobs",
        "produces": [
          "application/json"
        ],
        "tags": [
          "StoreManager"
        ],
        "operationId": "visits",
        "parameters": [
          {
            "description": "End Date",
            "name": "enddate",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success response",
            "schema": {
              "$ref": "#/definitions/VisitResponse"
            }
          },
          "400": {
            "description": "error response",
            "schema": {
              "$ref": "#/definitions/VisitResponse"
            }
          },
          "500": {
            "description": "error response",
            "schema": {
              "$ref": "#/definitions/VisitResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "StatusError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        },
        "store_id": {
          "type": "string",
          "x-go-name": "StoreID"
        }
      },
      "x-go-package": "storemanager/pkg/model"
    },
    "StatusResponse": {
      "description": "StatusResponse is model for submit response",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StatusError"
          },
          "x-go-name": "Errors"
        },
        "job_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "JobID"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "storemanager/pkg/model"
    },
    "SubmitRequest": {
      "description": "SubmitRequest is model for submit request",
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Count"
        },
        "visits": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Visit"
          },
          "x-go-name": "Visits"
        }
      },
      "x-go-package": "storemanager/pkg/model"
    },
    "SubmitResponse": {
      "description": "SubmitResponse is model for submit response",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        },
        "job_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "JobID"
        }
      },
      "x-go-package": "storemanager/pkg/model"
    },
    "Visit": {
      "type": "object",
      "properties": {
        "image_url": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "ImageURLs"
        },
        "store_id": {
          "type": "string",
          "x-go-name": "StoreID"
        },
        "visit_time": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "VisitTime"
        }
      },
      "x-go-package": "storemanager/pkg/model"
    },
    "VisitData": {
      "type": "object",
      "properties": {
        "date": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "Date"
        },
        "perimeter": {
          "type": "number",
          "format": "double",
          "x-go-name": "Perimeter"
        }
      },
      "x-go-package": "storemanager/pkg/model"
    },
    "VisitResponse": {
      "description": "VisitResponse is model for visit response",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        },
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/VisitResult"
          },
          "x-go-name": "Results"
        }
      },
      "x-go-package": "storemanager/pkg/model"
    },
    "VisitResult": {
      "type": "object",
      "properties": {
        "area": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Area"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/VisitData"
          },
          "x-go-name": "Datas"
        },
        "store_id": {
          "type": "string",
          "x-go-name": "StoreID"
        },
        "store_name": {
          "type": "string",
          "x-go-name": "StoreName"
        }
      },
      "x-go-package": "storemanager/pkg/model"
    }
  }
}